backend/requirements.txt
# Dawai Yaad Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
httpx==0.28.1
celery[redis]==5.4.0
redis==5.2.1
minio==7.2.12
firebase-admin==6.6.0
weasyprint==63.1
python-dateutil==2.9.0.post0
phonenumbers==8.13.52
pytest==8.3.4
pytest-asyncio==0.24.0
